some bug fix and refuctoring
