+++ yo/app/views/zeppelinDashboard.html
@@ -82,7 +82,7 @@
 
                         </div>
                         <hr>
-                        <p style="line-height: 1.02; margin-top: 10px"><span class="glyphicon glyphicon-info-sign" ></span>  If CPU quota is enforced starting too many interpreters can eat up your quota. Running a paragraph in a notebook will automatically start the necessary interpreters for that job.     </p>
+                        <p style="line-height: 1.02; margin-top: 10px"><span class="glyphicon glyphicon-info-sign" ></span>  Running a paragraph in a notebook will automatically start the necessary interpreters for that job.  </p>
                     </div>
                 </div>
             </div>
